+++ libsierp/include/sierp/sierp_point_list.h
@@ -17,7 +17,7 @@ int sierp_point_list_is_full(const SIERP_POINT_LIST *point_list);
 void sierp_point_list_size_set(SIERP_POINT_LIST *point_list, int newsize);
 int sierp_point_list_size_get(const SIERP_POINT_LIST *point_list);
 int sierp_point_list_push_point(SIERP_POINT_LIST *point_list, const SIERP_POINT *point);
-int sierp_point_list_push_int(SIERP_POINT_LIST *point_list, int x, int y);
+int sierp_point_list_push_int(SIERP_POINT_LIST *point_list, double x, double y);
 const SIERP_POINT *sierp_point_list_pop(SIERP_POINT_LIST *point_list);
 const SIERP_POINT *sierp_point_list_get_index(const SIERP_POINT_LIST *point_list, int index);
